requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.6
django-filter==24.2
django-phonenumber-field==7.3.0
django-rest-knox==4.2.0
djangorestframework==3.15.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
idna==3.7
Markdown==3.6
msgpack==1.0.8
oauthlib==3.2.2
phonenumbers==8.13.37
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyotp==2.9.0
redis==5.0.4
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
sorl-thumbnail==12.10.0
sqlparse==0.5.0
typing_extensions==4.12.0
urllib3==2.2.1
